replace in draw
